+++ Tests/SentryTests/TestClient.swift
@@ -2,7 +2,6 @@ import Foundation
 
 class TestClient: SentryClient {
     let sentryFileManager: SentryFileManager
-    let queue = DispatchQueue(label: "TestClient", attributes: .concurrent)
 
     override init?(options: Options) {
         sentryFileManager = try! SentryFileManager(options: options, andCurrentDateProvider: TestCurrentDateProvider())
@@ -83,16 +82,17 @@ class TestClient: SentryClient {
         captureExceptionWithScopeInvocations.record((exception, scope))
         return SentryId()
     }
-    
-    var captureErrorWithSessionInvocations = Invocations<(error: Error, session: SentrySession, scope: Scope)>()
-    override func captureError(_ error: Error, with session: SentrySession, with scope: Scope) -> SentryId {
-        captureErrorWithSessionInvocations.record((error, session, scope))
+
+    var callSessionBlockWithIncrementSessionErrors = true
+    var captureErrorWithSessionInvocations = Invocations<(error: Error, session: SentrySession?, scope: Scope)>()
+    override func captureError(_ error: Error, with scope: Scope, incrementSessionErrors sessionBlock: @escaping () -> SentrySession) -> SentryId {
+        captureErrorWithSessionInvocations.record((error, callSessionBlockWithIncrementSessionErrors ? sessionBlock() : nil, scope))
         return SentryId()
     }
     
-    var captureExceptionWithSessionInvocations = Invocations<(exception: NSException, session: SentrySession, scope: Scope)>()
-    override func capture(_ exception: NSException, with session: SentrySession, with scope: Scope) -> SentryId {
-        captureExceptionWithSessionInvocations.record((exception, session, scope))
+    var captureExceptionWithSessionInvocations = Invocations<(exception: NSException, session: SentrySession?, scope: Scope)>()
+    override func capture(_ exception: NSException, with scope: Scope, incrementSessionErrors sessionBlock: @escaping () -> SentrySession) -> SentryId {
+        captureExceptionWithSessionInvocations.record((exception, callSessionBlockWithIncrementSessionErrors ? sessionBlock() : nil, scope))
         return SentryId()
     }
     
@@ -129,9 +129,9 @@ class TestClient: SentryClient {
         recordLostEvents.record((category, reason))
     }
     
-    var flushInvoctions = Invocations<TimeInterval>()
+    var flushInvocations = Invocations<TimeInterval>()
     override func flush(timeout: TimeInterval) {
-        flushInvoctions.record(timeout)
+        flushInvocations.record(timeout)
     }
 }